+++ pkg/engine/plan.go
@@ -155,7 +155,7 @@ func plan(ctx *Context, info *planContext, opts planOptions, dryRun bool) (*plan
 	}
 
 	// Generate a plan; this API handles all interesting cases (create, update, delete).
-	plan, err := deploy.NewPlan(plugctx, target, target.Snapshot, source, analyzers, dryRun)
+	plan, err := deploy.NewPlan(plugctx, target, target.Snapshot, source, analyzers, dryRun, ctx.BackendClient)
 	if err != nil {
 		return nil, err
 	}